+++ packages/terra-core-docs/src/terra-dev-site/doc/search-field/Examples.2/SearchFieldInvalid.jsx
@@ -51,6 +51,8 @@ const SearchFieldInvalid = () => {
         value={text}
         inputRefCallback={(ref) => { inputRef.current = ref; }}
         searchDelay={500}
+        isLabelVisible
+        groupName="Find a clinic"
       />
       <div id="search-callback-text" className={isInvalid ? cx('error-text') : ''}>
         {message}