+++ packages/terra-core-docs/src/terra-dev-site/test/search-field/SearchFieldDisabledWithValue.test.jsx
@@ -1,4 +1,4 @@
 import React from 'react';
 import SearchField from 'terra-search-field';
 
-export default () => <SearchField id="searchfield" value="Value" isDisabled />;
+export default () => <SearchField id="searchfield" value="Value" isDisabled isLabelVisible groupName="Find a clinic" />;