+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 * Added
+
   * Added error message for `terra-form-tag-select` invalid example.
   * Added visual lebel for `terra-form-tag-select` example.
+  * Added visual labels for `terra-search-field` examples. 
 
 * Updated
   * Updated Draggable list item example for selection state in `terra-list`.